Refactor: build_eval_windows, moved to WFUtilities
